Calculate common space scores

The common space function extracts latent dimensions from a collection of roll call scaling results using the procedure outlined in Poole (1998) and Poole (2005) .

common_space(nom_list, id = NULL, dims = 2, minscale = 5, polarity = NULL)

Arguments

nom_list

A list of nomObject results from the wnominate package. The number of estimated dimensions in each nomObject should be greater than or equal to dims.

id

Column name in the nomObjects' legislators data frames providing a unique legislator ID. If not specified row names will be used.

dims

The number of dimensions to estimate.

minscale

The minimum number of W-NOMINATE scores a legislator must have to be used for the latent dimension estimates. Common space scores for legislators with fewer than minscale scores will be derived from the estimated dimensions.

polarity

A vector of ID's of length dims of legislators who should have relatively positive scores on the corresponding dimension. If only one ID is provided, it will be recycled for all dimensions.

Value

A list of class common space containing:

legislators A data frame of legislator coordinates

blackbox A list of blackbox results

nom_list The provided list of wnominate results

References

Poole KT (1998). “Recovering a Basic Space From a Set of Issue Scales.” American Journal of Political Science, 42(3), 954--993. ISSN 0092-5853, doi:10.2307/2991737 , https://www.jstor.org/stable/2991737.

Poole KT (2005). Spatial Models of Parliamentary Voting, Analytical Methods for Social Research. Cambridge University Press. doi:10.1017/CBO9780511614644 .

See also