New Hampshire Senate Roll Calls

New Hampshire Senate roll call votes from 1999 to 2019.

nhsenate

Format

A list of rollcall objects from the pscl package

Source

New Hampshire General Court

Examples

data(nhsenate)
summary(nhsenate[[1]])
#> 
#> Summary of rollcall object nhsenate[[1]] 
#> 
#> Description:	 Roll call votes from the New Hampshire Senate (1999) 
#> Source:		 http://gencourt.state.nh.us/nhgcrollcalls/ 
#> 
#> Number of Legislators:		 25
#> Number of Roll Call Votes:	 118
#> 
#> 
#> Using the following codes to represent roll call votes:
#> Yea:		 1 
#> Nay:		 0 
#> Abstentions:	 NA 
#> Not In Legislature:	 9 
#> 
#> Party Composition:
#>   Democrat Republican 
#>         13         12 
#> 
#> Vote Summary:
#>                Count Percent
#> 0 (nay)         1049    36.8
#> 1 (yea)         1667    58.6
#> 9 (notInLegis)   131     4.6
#> 
#> Use summary(nhsenate[[1]],verbose=TRUE) for more detailed information.